test(app): cover root error details without mounting
Move caught-value formatting into a pure module, cover AggregateError fields without duplicate output, and remove the component-mounted test.
